+++ src/main/java/phoupraw/mcmod/infinite_fluid_bucket/mixin/minecraft/MIngredient.java
@@ -1,19 +1,21 @@
 package phoupraw.mcmod.infinite_fluid_bucket.mixin.minecraft;
 
-import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
-import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.recipe.Ingredient;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import phoupraw.mcmod.infinite_fluid_bucket.dispenser.InfEmptyBucketBehavior;
 import phoupraw.mcmod.infinite_fluid_bucket.dispenser.InfGlassBottleBehavior;
 
-@Mixin(Ingredient.class)
+@Mixin(value = Ingredient.class, priority = 1100)
+//@Debug(export = true)
 abstract class MIngredient {
-    @WrapOperation(method = "test(Lnet/minecraft/item/ItemStack;)Z", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
-    private boolean checkInf(ItemStack instance, Item item, Operation<Boolean> original, ItemStack itemStack) {
-        return original.call(instance, item) && !InfEmptyBucketBehavior.isInf(itemStack) && !InfGlassBottleBehavior.isInf(itemStack);
+    @Inject(method = "test(Lnet/minecraft/item/ItemStack;)Z", at = @At("HEAD"), cancellable = true)
+    private void checkInf(ItemStack itemStack, CallbackInfoReturnable<Boolean> cir) {
+        if (InfEmptyBucketBehavior.isInf(itemStack) || InfGlassBottleBehavior.isInf(itemStack)) {
+            cir.setReturnValue(false);
+        }
     }
 }